requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.13.13

# Tools
django-extensions==3.2.1

# Testing
mypy==1.14.1
pytest== 6.2.4
pytest-sugar==0.9.4
pytest-django==4.3.0
factory-boy==2.11.1

# Code quality
flake8==3.6.0
